US-372_Task-374_Move Hamburger to separate class file and fix unauthenticated access bug
